upload: move page to views dir
